Fix unsigned/signed warnings in MSVC

+++ 3rdparty/OpenEXR/IlmImf/ImfOptimizedPixelReading.h
@@ -302,7 +302,7 @@ void writeToRGBASSETemplate
      __m128i*& writePtrSSE,
      const size_t& lPixelsToCopySSE)
 {
-    for (int i = 0; i < lPixelsToCopySSE; ++i)
+    for (size_t i = 0; i < lPixelsToCopySSE; ++i)
     {
         __m128i redRegister   = loadSSE<READ_PTR_ALIGNED> (readPtrSSERed);
         __m128i greenRegister = loadSSE<READ_PTR_ALIGNED> (readPtrSSEGreen);
@@ -359,7 +359,7 @@ void writeToRGBANormal (unsigned short*& readPtrRed,
                         unsigned short*& writePtr,
                         const size_t& lPixelsToCopy)
 {
-    for (int i = 0; i < lPixelsToCopy; ++i)
+    for (size_t i = 0; i < lPixelsToCopy; ++i)
     {
         *(writePtr++) = *(readPtrRed++);
         *(writePtr++) = *(readPtrGreen++);
@@ -461,7 +461,7 @@ writeToRGBAFillASSETemplate (__m128i*& readPtrSSERed,
                                                       alphaFillValue,
                                                       alphaFillValue);
 
-    for (int pixelCounter = 0; pixelCounter < pixelsToCopySSE; ++pixelCounter)
+    for (size_t pixelCounter = 0; pixelCounter < pixelsToCopySSE; ++pixelCounter)
     {
         __m128i redRegister   = loadSSE<READ_PTR_ALIGNED> (readPtrSSERed);
         __m128i greenRegister = loadSSE<READ_PTR_ALIGNED> (readPtrSSEGreen);
@@ -519,7 +519,7 @@ writeToRGBAFillANormal (unsigned short*& readPtrRed,
                         unsigned short*& writePtr,
                         const size_t& pixelsToCopy)
 {
-    for (int i = 0; i < pixelsToCopy; ++i)
+    for (size_t i = 0; i < pixelsToCopy; ++i)
     {
         *(writePtr++) = *(readPtrRed++);
         *(writePtr++) = *(readPtrGreen++);
@@ -613,7 +613,7 @@ writeToRGBSSETemplate (__m128i*& readPtrSSERed,
                        const size_t& pixelsToCopySSE)
 {
 
-    for (int pixelCounter = 0; pixelCounter < pixelsToCopySSE; ++pixelCounter)
+    for (size_t pixelCounter = 0; pixelCounter < pixelsToCopySSE; ++pixelCounter)
     {
         //
         // Need to shuffle and unpack pointers to obtain my first register
@@ -723,7 +723,7 @@ writeToRGBNormal (unsigned short*& readPtrRed,
                   unsigned short*& writePtr,
                   const size_t& pixelsToCopy)
 {
-    for (int i = 0; i < pixelsToCopy; ++i)
+    for (size_t i = 0; i < pixelsToCopy; ++i)
     {
         *(writePtr++) = *(readPtrRed++);
         *(writePtr++) = *(readPtrGreen++);
